+++ src/appleseed/renderer/kernel/lighting/sppm/sppmparameters.h
@@ -70,7 +70,7 @@ struct SPPMParameters
     const float                 m_initial_radius_percents;              // initial lookup radius as a percentage of the scene diameter
     const float                 m_alpha;                                // radius shrinking control
     const size_t                m_max_photons_per_estimate;             // maximum number of photons per density estimation
-    const double                m_dl_light_sample_count;                // number of light samples used to estimate direct illumination in ray traced mode
+    const float                 m_dl_light_sample_count;                // number of light samples used to estimate direct illumination in ray traced mode
     float                       m_rcp_dl_light_sample_count;
 
     const bool                  m_view_photons;                         // debug mode to visualize the photons
