+++ src/appleseed/renderer/meta/tests/test_tracer.cpp
@@ -275,7 +275,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double transmission;
+        float transmission;
         const ShadingPoint& shading_point =
             tracer.trace(
                 Vector3d(0.0),
@@ -286,7 +286,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 transmission);
 
         EXPECT_FALSE(shading_point.hit());
-        EXPECT_EQ(1.0, transmission);
+        EXPECT_EQ(1.0f, transmission);
     }
 
     TEST_CASE_F(Trace_QuickVariant_GivenNoOccluder, Fixture<EmptyScene>)
@@ -300,7 +300,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        const double transmission =
+        const float transmission =
             tracer.trace(
                 Vector3d(0.0),
                 Vector3d(1.0, 0.0, 0.0),
@@ -308,7 +308,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 VisibilityFlags::ShadowRay,
                 0);
 
-        EXPECT_EQ(1.0, transmission);
+        EXPECT_EQ(1.0f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_GivenNoOccluder, Fixture<EmptyScene>)
@@ -322,7 +322,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double transmission;
+        float transmission;
         const ShadingPoint& shading_point =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
@@ -333,7 +333,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 transmission);
 
         EXPECT_FALSE(shading_point.hit());
-        EXPECT_EQ(1.0, transmission);
+        EXPECT_EQ(1.0f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_QuickVariant_GivenNoOccluder, Fixture<EmptyScene>)
@@ -347,7 +347,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        const double transmission =
+        const float transmission =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
                 Vector3d(5.0, 0.0, 0.0),
@@ -355,7 +355,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 VisibilityFlags::ShadowRay,
                 0);
 
-        EXPECT_EQ(1.0, transmission);
+        EXPECT_EQ(1.0f, transmission);
     }
 
     struct SceneWithSingleOpaqueOccluder
@@ -378,7 +378,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double transmission;
+        float transmission;
         const ShadingPoint& shading_point =
             tracer.trace(
                 Vector3d(0.0, 0.0, 0.0),
@@ -390,7 +390,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 
         ASSERT_TRUE(shading_point.hit());
         EXPECT_FEQ(Vector3d(2.0, 0.0, 0.0), shading_point.get_point());
-        EXPECT_EQ(1.0, transmission);
+        EXPECT_EQ(1.0f, transmission);
     }
 
     TEST_CASE_F(Trace_QuickVariant_GivenSingleOpaqueOccluder, Fixture<SceneWithSingleOpaqueOccluder>)
@@ -404,7 +404,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        const double transmission =
+        const float transmission =
             tracer.trace(
                 Vector3d(0.0, 0.0, 0.0),
                 Vector3d(1.0, 0.0, 0.0),
@@ -412,7 +412,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 VisibilityFlags::ShadowRay,
                 0);
 
-        EXPECT_EQ(0.0, transmission);
+        EXPECT_EQ(0.0f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_GivenSingleOpaqueOccluder, Fixture<SceneWithSingleOpaqueOccluder>)
@@ -426,7 +426,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double transmission;
+        float transmission;
         const ShadingPoint& shading_point =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
@@ -438,7 +438,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 
         ASSERT_TRUE(shading_point.hit());
         EXPECT_FEQ(Vector3d(2.0, 0.0, 0.0), shading_point.get_point());
-        EXPECT_EQ(1.0, transmission);
+        EXPECT_EQ(1.0f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_QuickVariant_GivenSingleOpaqueOccluder, Fixture<SceneWithSingleOpaqueOccluder>)
@@ -452,7 +452,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        const double transmission =
+        const float transmission =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
                 Vector3d(5.0, 0.0, 0.0),
@@ -460,7 +460,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 VisibilityFlags::ShadowRay,
                 0);
 
-        EXPECT_EQ(0.0, transmission);
+        EXPECT_EQ(0.0f, transmission);
     }
 
     struct SceneWithSingleTransparentOccluder
@@ -483,7 +483,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double transmission;
+        float transmission;
         const ShadingPoint& shading_point =
             tracer.trace(
                 Vector3d(0.0, 0.0, 0.0),
@@ -494,7 +494,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 transmission);
 
         ASSERT_FALSE(shading_point.hit());
-        EXPECT_FEQ(0.5, transmission);
+        EXPECT_FEQ(0.5f, transmission);
     }
 
     TEST_CASE_F(Trace_QuickVariant_GivenSingleTransparentOccluder, Fixture<SceneWithSingleTransparentOccluder>)
@@ -508,7 +508,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        const double transmission =
+        const float transmission =
             tracer.trace(
                 Vector3d(0.0, 0.0, 0.0),
                 Vector3d(1.0, 0.0, 0.0),
@@ -516,7 +516,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 VisibilityFlags::ShadowRay,
                 0);
 
-        EXPECT_FEQ(0.5, transmission);
+        EXPECT_FEQ(0.5f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_GivenSingleTransparentOccluder, Fixture<SceneWithSingleTransparentOccluder>)
@@ -530,7 +530,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double transmission;
+        float transmission;
         const ShadingPoint& shading_point =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
@@ -541,7 +541,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 transmission);
 
         ASSERT_FALSE(shading_point.hit());
-        EXPECT_FEQ(0.5, transmission);
+        EXPECT_FEQ(0.5f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_QuickVariant_GivenSingleTransparentOccluder, Fixture<SceneWithSingleTransparentOccluder>)
@@ -555,7 +555,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        const double transmission =
+        const float transmission =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
                 Vector3d(5.0, 0.0, 0.0),
@@ -563,7 +563,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 VisibilityFlags::ShadowRay,
                 0);
 
-        EXPECT_FEQ(0.5, transmission);
+        EXPECT_FEQ(0.5f, transmission);
     }
 
     struct SceneWithTransparentThenOpaqueOccluders
@@ -587,7 +587,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double transmission;
+        float transmission;
         const ShadingPoint& shading_point =
             tracer.trace(
                 Vector3d(0.0, 0.0, 0.0),
@@ -599,7 +599,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 
         ASSERT_TRUE(shading_point.hit());
         EXPECT_FEQ(Vector3d(4.0, 0.0, 0.0), shading_point.get_point());
-        EXPECT_FEQ(0.5, transmission);
+        EXPECT_FEQ(0.5f, transmission);
     }
 
     TEST_CASE_F(Trace_QuickVariant_GivenTransparentThenOpaqueOccluders, Fixture<SceneWithTransparentThenOpaqueOccluders>)
@@ -613,7 +613,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        const double transmission =
+        const float transmission =
             tracer.trace(
                 Vector3d(0.0, 0.0, 0.0),
                 Vector3d(1.0, 0.0, 0.0),
@@ -621,7 +621,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 VisibilityFlags::ShadowRay,
                 0);
 
-        EXPECT_FEQ(0.0, transmission);
+        EXPECT_FEQ(0.0f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_GivenTransparentThenOpaqueOccluders_GivenTargetPastOpaqueOccluder, Fixture<SceneWithTransparentThenOpaqueOccluders>)
@@ -635,7 +635,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double transmission;
+        float transmission;
         const ShadingPoint& shading_point =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
@@ -647,7 +647,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 
         ASSERT_TRUE(shading_point.hit());
         EXPECT_FEQ(Vector3d(4.0, 0.0, 0.0), shading_point.get_point());
-        EXPECT_FEQ(0.5, transmission);
+        EXPECT_FEQ(0.5f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_QuickVariant_GivenTransparentThenOpaqueOccluders_GivenTargetPastOpaqueOccluder, Fixture<SceneWithTransparentThenOpaqueOccluders>)
@@ -661,7 +661,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        const double transmission =
+        const float transmission =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
                 Vector3d(5.0, 0.0, 0.0),
@@ -669,7 +669,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 VisibilityFlags::ShadowRay,
                 0);
 
-        EXPECT_FEQ(0.0, transmission);
+        EXPECT_FEQ(0.0f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_GivenTransparentThenOpaqueOccluders_GivenTargetOnOpaqueOccluder, Fixture<SceneWithTransparentThenOpaqueOccluders>)
@@ -683,7 +683,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double transmission;
+        float transmission;
         const ShadingPoint& shading_point =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
@@ -694,7 +694,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 transmission);
 
         ASSERT_FALSE(shading_point.hit());
-        EXPECT_FEQ(0.5, transmission);
+        EXPECT_FEQ(0.5f, transmission);
     }
 
     TEST_CASE_F(TraceBetween_QuickVariant_GivenTransparentThenOpaqueOccluders_GivenTargetOnOpaqueOccluder, Fixture<SceneWithTransparentThenOpaqueOccluders>)
@@ -708,7 +708,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        const double transmission =
+        const float transmission =
             tracer.trace_between(
                 Vector3d(0.0, 0.0, 0.0),
                 Vector3d(4.0, 0.0, 0.0),
@@ -716,7 +716,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
                 VisibilityFlags::ShadowRay,
                 0);
 
-        EXPECT_FEQ(0.5, transmission);
+        EXPECT_FEQ(0.5f, transmission);
     }
 
     struct SceneWithTwoOpaqueOccluders
@@ -740,7 +740,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double parent_transmission;
+        float parent_transmission;
         const ShadingPoint& parent_shading_point =
             parent_tracer.trace(
                 Vector3d(0.0, 0.0, 0.0),
@@ -761,13 +761,13 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
             , *m_shading_group_exec
 #endif
             );
-        const double transmission =
+        const float transmission =
             tracer.trace_between(
                 parent_shading_point,
                 Vector3d(4.0, 0.0, 0.0),
                 VisibilityFlags::ShadowRay);
 
-        EXPECT_EQ(1.0, transmission);
+        EXPECT_EQ(1.0f, transmission);
     }
 
     struct SceneWithTwoOpaqueOccludersAndScaledAssemblyInstance
@@ -776,7 +776,8 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
         SceneWithTwoOpaqueOccludersAndScaledAssemblyInstance()
         {
             m_assembly_instance->transform_sequence().set_transform(
-                0.0, Transformd::from_local_to_parent(
+                0.0f,
+                Transformd::from_local_to_parent(
                     Matrix4d::make_scaling(Vector3d(0.5))));
         }
     };
@@ -792,7 +793,7 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
 #endif
             );
 
-        double parent_transmission;
+        float parent_transmission;
         const ShadingPoint& parent_shading_point =
             parent_tracer.trace(
                 Vector3d(0.0, 0.0, 0.0),
@@ -813,12 +814,12 @@ TEST_SUITE(Renderer_Kernel_Lighting_Tracer)
             , *m_shading_group_exec
 #endif
             );
-        const double transmission =
+        const float transmission =
             tracer.trace_between(
                 parent_shading_point,
                 Vector3d(2.0, 0.0, 0.0),
                 VisibilityFlags::ShadowRay);
 
-        EXPECT_EQ(1.0, transmission);
+        EXPECT_EQ(1.0f, transmission);
     }
 }
