+++ src/appleseed/renderer/modeling/environmentedf/environmentedf.cpp
@@ -77,7 +77,7 @@ bool EnvironmentEDF::on_frame_begin(
     bool warned = false;
     for (size_t i = 0, e = m_transform_sequence.size(); i < e; ++i)
     {
-        double time;
+        float time;
         Transformd transform;
         m_transform_sequence.get_transform(i, time, transform);
